+++ src/components/GasCustomize/Ethereum/GasSelect.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { getRecoil } from 'recoil-nexus';
-import { EVM_MINIMUM_GAS_LIMIT, EVM_MINIMUM_GAS_PRICE } from '../../../config/StaticConfig';
+import { EVM_MINIMUM_GAS_LIMIT, EVM_MINIMUM_GAS_PRICE, ThemeColor } from '../../../config/StaticConfig';
 import { useAnalytics } from '../../../hooks/useAnalytics';
 import { useMarketPrice } from '../../../hooks/useMarketPrice';
 import { UserAsset, UserAssetType } from '../../../models/UserAsset';
@@ -169,7 +169,7 @@ export const GasStepSelectEthereum = ({ asset, onChange }: IGasStepSelectEthereu
           style={{ cursor: 'pointer' }}
           title={t('sending-crypto-on-blockchain-requires-confirmation')}
         >
-          <ExclamationCircleOutlined style={{ color: '#1199fa', cursor: 'pointer' }} />
+          <ExclamationCircleOutlined style={{ color: ThemeColor.BLUE, cursor: 'pointer' }} />
         </Tooltip>
       </div>
       {isUsingCustomGas ? (