+++ src/components/NoticeDisclaimer/NoticeDisclaimer.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Layout } from 'antd';
 import { ExclamationCircleOutlined } from '@ant-design/icons';
 import './NoticeDisclaimer.less';
+import { ThemeColor } from '../../config/StaticConfig';
 
 const { Content, Sider } = Layout;
 
@@ -10,7 +11,7 @@ const NoticeDisclaimer: React.FC = props => {
     <div className="notice-disclaimer">
       <Layout>
         <Sider width="20px">
-          <ExclamationCircleOutlined style={{ color: '#1199fa' }} />
+          <ExclamationCircleOutlined style={{ color: ThemeColor.BLUE }} />
         </Sider>
         <Content>{props.children}</Content>
       </Layout>