+++ src/pages/assets/components/FormSend.tsx
@@ -31,7 +31,7 @@ import {
   fromScientificNotation,
   getCurrentMinAssetAmount,
 } from '../../../utils/NumberUtils';
-import { FIXED_DEFAULT_FEE, SUPPORTED_CURRENCY } from '../../../config/StaticConfig';
+import { FIXED_DEFAULT_FEE, SUPPORTED_CURRENCY, ThemeColor } from '../../../config/StaticConfig';
 import { detectConditionsError, LEDGER_WALLET_TYPE } from '../../../service/LedgerService';
 import {
   AnalyticsActions,
@@ -46,6 +46,8 @@ import { AddressBookContact } from '../../../models/AddressBook';
 import { useLedgerStatus } from '../../../hooks/useLedgerStatus';
 import GasStepSelect from '../../../components/GasCustomize/GasConfig';
 import GasInfo from '../../../components/GasCustomize/GasInfo';
+import { ExclamationCircleOutlined } from '@ant-design/icons';
+import { checkIsDefaultUserAssetConfig } from '../../../config/StaticAssets';
 
 const layout = {};
 const tailLayout = {};
@@ -422,6 +424,15 @@ const FormSend: React.FC<FormSendProps> = props => {
                   <div>--</div>
                 )}
             </div>
+            {!checkIsDefaultUserAssetConfig(walletAsset, currentSession) ? 
+              <div className="item">
+                <ExclamationCircleOutlined
+                  style={{ color: ThemeColor.RED, marginRight: '5px', fontSize: '16px' }}
+                /> 
+                {t('assets.defaultConfig.tooltip')}
+              </div>
+              : ''
+            }
           </>
         </ModalPopup>